feat(desktop): shell-shaped skeleton for the connection gate
While connecting (plus a 10s startup grace on managed hosts) the gate
renders a static skeleton mirroring the shell instead of gate text, so
the daemon's ~250ms boot reads as the app loading; the failure copy
appears only once genuinely errored. The grace moves out of the shared
ConnectionState — it is the desktop fallback's branch condition now.
