+++ apps/desktop/src/renderer/src/app.tsx
@@ -1,6 +1,7 @@
 import {
   ConnectionState,
   createDaemonTransport,
+  useWorkbenchRuntimeStatus,
   Workbench,
   WorkbenchAppProviders,
   WorkbenchProviders,
@@ -11,6 +12,7 @@ import { useState } from 'react';
 import { systemBridge } from './ipc';
 import { SettingsView } from './settings/settings-view';
 import { useDesktopSettingsStore } from './settings/store';
+import { ConnectionSkeleton } from './shell/connection-skeleton';
 import { DesktopWorkbenchShell } from './shell/desktop-workbench-shell';
 
 export function DesktopApp(): React.ReactNode {
@@ -52,16 +54,32 @@ function DaemonConnection({
 }
 
 /**
- * Desktop connection gate: the shared `ConnectionState` plus endpoint rediscovery. The transport
- * retries a fixed URL, but the daemon port-hunts (apps/daemon/src/runtime.ts) — a daemon that
- * (re)started on another port is only reachable by re-resolving discovery. Main pushes a
- * runtime-file change event (fs.watch on ~/.linkcode); mounted only while the gate is closed, so
- * the subscription ends as soon as the transport connects.
+ * A supervised daemon needs a beat to boot (fork + engine + listener bind, ~250ms measured);
+ * early dial failures within this window are startup, not an outage — keep the skeleton up.
+ */
+const MANAGED_STARTUP_GRACE_MS = 10000;
+
+/**
+ * Desktop connection gate: a shell-shaped skeleton while connecting (plus a startup grace window
+ * on managed hosts), the shared `ConnectionState` once genuinely errored — and endpoint
+ * rediscovery throughout. The transport retries a fixed URL, but the daemon port-hunts
+ * (apps/daemon/src/runtime.ts) — a daemon that (re)started on another port is only reachable by
+ * re-resolving discovery. Main pushes a runtime-file change event (fs.watch on ~/.linkcode);
+ * mounted only while the gate is closed, so the subscription ends once the transport connects.
  */
 function DesktopConnectionFallback({ daemonUrl }: { daemonUrl: string }): React.ReactNode {
+  const status = useWorkbenchRuntimeStatus();
   const hasOverride = useDesktopSettingsStore((state) => state.daemonUrlOverride !== null);
   const adoptDiscoveredUrl = useDesktopSettingsStore((state) => state.adoptDiscoveredUrl);
   const managed = useDaemonIsManaged();
+
+  const [withinStartupGrace, setWithinStartupGrace] = useState(true);
+  useAbortableEffect((signal) => {
+    const timer = setTimeout(() => {
+      if (!signal.aborted) setWithinStartupGrace(false);
+    }, MANAGED_STARTUP_GRACE_MS);
+    signal.addEventListener('abort', () => clearTimeout(timer));
+  }, []);
 
   useAbortableEffect(
     (signal) => {
@@ -77,6 +95,7 @@ function DesktopConnectionFallback({ daemonUrl }: { daemonUrl: string }): React.
     [hasOverride, daemonUrl, adoptDiscoveredUrl],
   );
 
+  if (status === 'connecting' || (managed && withinStartupGrace)) return <ConnectionSkeleton />;
   return <ConnectionState daemonUrl={daemonUrl} managedHost={managed} />;
 }
