+++ packages/workbench/src/app/connection-state.tsx
@@ -1,15 +1,7 @@
 import { DAEMON_DEFAULT_URL } from '@linkcode/schema';
 import { Button } from 'coss-ui/components/button';
-import { useAbortableEffect } from 'foxact/use-abortable-effect';
-import { useState } from 'react';
 import { useTranslations } from 'use-intl';
 import { useWorkbenchRuntimeRetry, useWorkbenchRuntimeStatus } from '../runtime/provider';
-
-/**
- * A supervised daemon needs a beat to boot (fork + engine + listener bind); early dial failures
- * within this window after the gate opens are startup, not an outage — keep showing "connecting".
- */
-const MANAGED_STARTUP_GRACE_MS = 10000;
 
 /** Default connection-gate fallback: shown while the transport connects or after it errored. */
 export function ConnectionState({
@@ -25,18 +17,10 @@ export function ConnectionState({
   const t = useTranslations('workbench.connection');
   const common = useTranslations('common');
 
-  const [withinStartupGrace, setWithinStartupGrace] = useState(true);
-  useAbortableEffect((signal) => {
-    const timer = setTimeout(() => {
-      if (!signal.aborted) setWithinStartupGrace(false);
-    }, MANAGED_STARTUP_GRACE_MS);
-    signal.addEventListener('abort', () => clearTimeout(timer));
-  }, []);
-
   return (
     <div className="flex h-full items-center justify-center p-8">
       <div className="max-w-md text-center">
-        {status === 'connecting' || (managedHost && withinStartupGrace) ? (
+        {status === 'connecting' ? (
           <p className="text-muted-foreground text-sm">{t('connecting')}</p>
         ) : (
           <div className="space-y-3">
